Separate EV3 code into own module
