+++ src/robots/ev3/CMakeLists.txt
@@ -1,0 +1,5 @@
+cmake_minimum_required(VERSION 3.1)
+include(../../../cmake/bob_robotics.cmake)
+BoB_module(SOURCES ev3.cc
+           BOB_MODULES robots
+           THIRD_PARTY ev3dev-lang-cpp)
